paruparuku-webapp/local-requirements.txt
Flask==2.1.3
Matplotlib==3.6.2
Keras==2.7.0
Numpy==1.22.4
Tensorflow==2.7.0
Pillow==9.2.0
flask-babel==2.0.0